Refactor posts API to tasks
